requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
django-import-export==0.4.5
postgres==2.2.1
psycopg2==2.6.1
